+++ ivy/array/extensions/activations.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithActivationsExtensions(abc.ABC):
+    pass
