+++ ivy/array/extensions/conversions.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithConversionsExtensions(abc.ABC):
+    pass
