+++ ivy/array/extensions/creation.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithCreationExtensions(abc.ABC):
+    pass
