+++ ivy/array/extensions/data_type.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithData_typeExtensions(abc.ABC):
+    pass
