+++ ivy/array/extensions/device.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithDeviceExtensions(abc.ABC):
+    pass
