+++ ivy/array/extensions/general.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithGeneralExtensions(abc.ABC):
+    pass
