+++ ivy/array/extensions/gradients.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithGradientsExtensions(abc.ABC):
+    pass
