+++ ivy/array/extensions/image.py
@@ -1,8 +1,6 @@
 # global
 import abc
 
-# ToDo: implement all methods here as public instance methods
-
 
 class ArrayWithImageExtension(abc.ABC):
     pass
