+++ ivy/array/extensions/linear_algebra.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithLinalgExtensions(abc.ABC):
+    pass
