+++ ivy/array/extensions/losses.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithLossesExtensions(abc.ABC):
+    pass
