+++ ivy/array/extensions/norms.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithNormsExtensions(abc.ABC):
+    pass
