+++ ivy/array/extensions/random.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithRandomExtensions(abc.ABC):
+    pass
