+++ ivy/array/extensions/searching.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithSearchingExtensions(abc.ABC):
+    pass
