+++ ivy/array/extensions/set.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithSetExtensions(abc.ABC):
+    pass
