+++ ivy/array/extensions/sorting.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithSortingExtensions(abc.ABC):
+    pass
