+++ ivy/array/extensions/utility.py
@@ -1,0 +1,6 @@
+# global
+import abc
+
+
+class ArrayWithUtilityExtensions(abc.ABC):
+    pass
